refactor(types): extract product tests to sibling file
